+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.5
+- Add bypass for the queryDeleted function for non CRDT databases.
+
+## 1.0.4
+- Added an 'inMemory' constructor to the database. This is useful for testing and development.
+
+## 1.0.3
+- Update Readme with migrations instructions.
+
 ## 1.0.2
 - Upgraded synchroflite to 0.1.1